I've consolidated the initialization file generation logic.
I moved functions from `packages/shadcn/src/commands/init/init-file-generation.ts` into `packages/shadcn/src/commands/init/index.ts`.

The following functions were moved:
- `writeProjectConfig`
- `generateProjectTypes`
- `updateManifestYaml`

This change centralizes the initialization logic within the `init` command's main file (`index.ts`) and removes the separate `init-file-generation.ts` file, simplifying the directory structure.

Imports in `index.ts` have been updated to use the moved functions directly, and the now-obsolete `init-file-generation.ts` file has been removed.
